requirements.txt
fastapi
uvicorn[standard]
pyporscheconnectapi
jinja2
python-multipart
itsdangerous
Pillow
